upsert logic moved to monitor_service
